+++ src/resources/views/pages/edit.blade.php
@@ -29,22 +29,12 @@
                         </div>
                     </div>
                 </div>
-                <div class="panel panel-default">
-                    <div class="panel-heading">
-                        <div class="panel-title">{{ Lang::get('redminportal::forms.category') }}</div>
-                    </div>
-                    <div class="panel-body">
-                        {!! Form::hidden('category_id', $page->category_id, array('id' => 'category_id')) !!}
-                        <ul class="redooor-hierarchy">
-                            <li>
-                                <a href="0"><span class="glyphicon glyphicon-chevron-right"></span> {{ Lang::get('redminportal::forms.no_category') }}</a>
-                            </li>
-                        @foreach ($categories as $item)
-                            <li>{!! $item->printCategory() !!}</li>
-                        @endforeach
-                        </ul>
-                    </div>
-                </div>
+                {{-- Load Select Category partial form --}}
+                @include('redminportal::partials.form-select-category', [
+                    'select_category_selected_name' => 'category_id',
+                    'select_category_selected_id' => $page->category_id,
+                    'select_category_categories' => $categories
+                ])
                 <div>
                     <div class="fileupload fileupload-new" data-provides="fileupload">
                       <div class="fileupload-preview thumbnail" style="width: 200px; height: 150px;"></div>
@@ -153,6 +143,7 @@
 @stop
 
 @section('footer')
+    @parent
     <script src="{{ URL::to('vendor/redooor/redminportal/js/bootstrap-fileupload.js') }}"></script>
     <script>
         !function ($) {
@@ -161,29 +152,6 @@
                 $('#lang-selector a').click(function (e) {
                     e.preventDefault();
                     $(this).tab('show');
-                });
-                // On load, check if previous category exists for error message
-                function checkCategory() {
-                    $selected_val = $('#category_id').val();
-                    if ($selected_val == '') {
-                        // Initialize No Parent
-                        $('#parent_id').val(0);
-                        $selected_val = '0';
-                    }
-                    $('.redooor-hierarchy a').each(function() {
-                        if ($(this).attr('href') == $selected_val) {
-                            $(this).addClass('active');
-                        }
-                    });
-                }
-                checkCategory();
-                // Change selected category
-                $(document).on('click', '.redooor-hierarchy a', function(e) {
-                    e.preventDefault();
-                    $selected = $(this).attr('href');
-                    $('#category_id').val($selected);
-                    $('.redooor-hierarchy a.active').removeClass('active');
-                    $(this).addClass('active');
                 });
             })
         }(window.jQuery);