+++ src/resources/views/pages/home.blade.php
@@ -27,6 +27,7 @@
                         <li>Add getTotaldiscount method to Order. (issue #142)</li>
                         <li>Improved Tinymce editor to match RedminPortal look and feel.</li>
                         <li>Updated Tinymce to version 4.3.2.</li>
+                        <li>Create Category Select reusable form (issue #130)</li>
                     </ol>
                     <h4>Bug fixes:</h4>
                     <ol>