+++ src/resources/views/products/edit-variant.blade.php
@@ -167,19 +167,13 @@
                         </div>
                     </div>
                 </div>
-                <div class="panel panel-default">
-                    <div class="panel-heading">
-                        <div class="panel-title">{{ Lang::get('redminportal::forms.category') }}</div>
-                    </div>
-                    <div class="panel-body">
-                        {!! Form::hidden('category_id', $product->category_id, array('id' => 'category_id')) !!}
-                        <ul class="redooor-hierarchy">
-                        @foreach ($categories as $item)
-                            <li>{!! $item->printCategory() !!}</li>
-                        @endforeach
-                        </ul>
-                    </div>
-                </div>
+                {{-- Load Select Category partial form --}}
+                @include('redminportal::partials.form-select-category', [
+                    'select_category_selected_name' => 'category_id',
+                    'select_category_selected_id' => $product->category_id,
+                    'select_category_categories' => $categories,
+                    'select_category_required_field' => true
+                ])
 		        <div>
                     <div class="fileupload fileupload-new" data-provides="fileupload">
                       <div class="fileupload-preview thumbnail" style="width: 200px; height: 150px;"></div>
@@ -210,6 +204,7 @@
 @stop
 
 @section('footer')
+    @parent
     <script src="{{ URL::to('vendor/redooor/redminportal/js/bootstrap-fileupload.js') }}"></script>
     <script>
         !function ($) {
@@ -219,26 +214,6 @@
                     $delete_url = $(this).attr('href');
                     $('#confirm-modal-proceed').attr('href', $delete_url);
                     $('#confirm-modal').modal('show');
-                });
-                // On load, check if previous category exists for error message
-                function checkCategory() {
-                    $selected_val = $('#category_id').val();
-                    if ($selected_val != '') {
-                        $('.redooor-hierarchy a').each(function() {
-                            if ($(this).attr('href') == $selected_val) {
-                                $(this).addClass('active');
-                            }
-                        });
-                    }
-                }
-                checkCategory();
-                // Change selected category
-                $(document).on('click', '.redooor-hierarchy a', function(e) {
-                    e.preventDefault();
-                    $selected = $(this).attr('href');
-                    $('#category_id').val($selected);
-                    $('.redooor-hierarchy a.active').removeClass('active');
-                    $(this).addClass('active');
                 });
             })
         }(window.jQuery);